refactor typed contract
